correcion d eproducts2

+++ src/main/java/com/clicknsweet/clicknsweet/exceptions/CartNotFoundException.java
@@ -1,7 +1,7 @@
 package com.clicknsweet.clicknsweet.exceptions;
 
 public class CartNotFoundException extends RuntimeException {
-  public CartNotFoundException(Integer id) {
+  public CartNotFoundException(Long id) {
     super("No se encontró el carrito " + id);
   }
   public CartNotFoundException(String by) {
